Improve `helper_command' config default
This option needs to capture the current oslo_config in some way, so it
can be reconstructed in the new privileged process (when using the
'sudo/rootwrap' method).  The previous version had a "$project"
placeholder default that didn't work in real usage (as expected).  This
new version generates a default value at run-time based on the values of
cfg.CONF.{config_file,config_dir}.

If the deployer has provided an explicit value for `helper_command' then
the new cfg.CONF logic is also ignored.

Note secure deployments will capture this command line in sudoers or
rootwrap filters, and it is up to sudo/rootwrap to verify that whatever
we generate here is secure and reasonable.

Note also that this and the surrounding code is ignored when using the
'fork' method.

Change-Id: I0d31bf24cac6c26f10b5d1eebaa8f475402f73d2