Remove unused py27 socketpair/makefile workaround
Python2.7 has a bug that means the sockets returned from socketpair()
are not wrapped in the usual socket.socket class, and contain a broken
makefile() method.

This was important a long time ago when privsep used file
operations (and sock.makefile()) for communication.  Privsep now uses
socket operations directly, and never calls makefile() on the
communication socket.

Further, the workaround breaks on py34 where the socket classes are
completely different (and don't have the original issue).  Well past
time for the workaround to be removed and forgotten.

Change-Id: Id6395b8497b9f5449893e17d07fc4b664ee041ce