Bring sanity to lower-constraints
This is all necessary to bump the lower-constraint on msgpack so we
can support msgpack 1.0 without version-specific logic in our code.

The reasons for the changes are as follows:

* greenlet and PyYAML bumped because the previous lower-constraints
  for those didn't install on python 3.7, which is a supported Python
  version.
* msgpack-python is removed. It was replaced by msgpack and there is
  no new enough release of it to satisfy the needed minimum version bump.
* oslo.serialization is bumped because the old version was pulling in
  msgpack-python, which overwrote our new needed version of msgpack.
* I went ahead and included the msgpack bump to 0.6.0 so we can move
  forward in the subsequent patch with supporting msgpack 1.0.0.

It could be argued that this should be included in the msgpack 1.0.0
change, but it ended up being complex enough that I thought it was
worth it to split it out.

Change-Id: I69dae65d3e0a40bb2304d74de078ab84fc778d58