Add except path with exception debug to send_recv
The related bug resulted when an exception occurred within the
future.result() call. This caused the finally block to be executed,
and therefore myid to be deleted from self.outstanding_msgs prior
to _reader_main() checking if the msgid not in self.outstanding_msgs.
This caused _reader_main() to raise an AssertionError because the
msgid was no longer in outstanding_msgs. This is a small step forward
to log a warning when this siutation occurs.

Related-Bug: #1927868
Change-Id: I2eed242e0c796b8a2aa3d1b21bd1da4c497f624d