Setup logging without fixing evenlet logging
Since [1], ``oslo_log.log.setup`` can be called without applying the
fix for eventlet in native threads [2]. This fix clashes with the
oslo.privsep logging handler that replaces the original one. This
handler is implemented to allow the sync between the daemon process
and the process making the privileged call.

Once the oslo.log library version is bumped to 5.0.2, the try clause
can be removed.

[1]https://review.opendev.org/c/openstack/oslo.log/+/864252
[2]https://review.opendev.org/c/openstack/oslo.log/+/852443

Closes-Bug: #1995514
Related-Bug: #1995091
Change-Id: I7a4c55228064cb2dd4f4a359cdd81fd288baaf68